made snapshots optional if $VHOST is not set

+++ features/support/snapshot_hooks.rb
@@ -1,19 +1,23 @@
 # snapshot books
 Before do |scenario|
     #take a snapshot 
-    step "I take a snapshot \"#{short_name(scenario.name)}\""
+    if ENV['VHOST']
+        step "I take a snapshot \"#{short_name(scenario.name)}\""
+    end
 end
 
 After do |scenario|
     #take a snapshot if the scenario fails
-    name = short_name(scenario.name)
-    if(scenario.failed?)
-        step "I take a snapshot \"#{name}_failed\""
-    else
-        #nothing happened, delete snapshot as we dont need it   
-        $sshout = ""
-        $sshout = `echo | ssh root@$VHOST rm -f $IMGDIR/#{name}.qcow2`
-        raise "Failed to remove snapshot" unless $?.success? 
+    if ENV['VHOST']
+        name = short_name(scenario.name)
+        if(scenario.failed?)
+            step "I take a snapshot \"#{name}_failed\""
+        else
+            #nothing happened, delete snapshot as we dont need it   
+            $sshout = ""
+            $sshout = `echo | ssh root@$VHOST rm -f $IMGDIR/#{name}.qcow2`
+            raise "Failed to remove snapshot" unless $?.success? 
+        end
     end
 end
